+++ stapi-pydantic/src/stapi_pydantic/order.py
@@ -129,6 +129,7 @@ class OrderCollection(_GeoJsonBase, Generic[T]):
     type: Literal["FeatureCollection"] = "FeatureCollection"
     features: list[Order[T]]
     links: list[Link] = Field(default_factory=list)
+    numberMatched: int | None = Field(default=None, exclude_if=lambda x: x is None)
 
     def __iter__(self) -> Iterator[Order[T]]:  # type: ignore [override]
         """iterate over features"""